# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User and Fish API",
    "description": "API documentation for managing users and fish",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/users/": {
      "post": {
        "summary": "Create a new user",
        "description": "Creates a new user with the provided details",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all users",
        "description": "Returns a list of all users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login a user",
        "description": "Authenticates a user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/peixes/{userId}": {
      "get": {
        "summary": "Get fish by user ID",
        "description": "Returns a list of fish associated with a specific user ID",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of fish",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Peixe"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Add a fish to a user",
        "description": "Adds a fish to the specified user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/PeixesUsuario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fish added successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/peixes/{id}": {
      "get": {
        "summary": "Get fish details by ID",
        "description": "Returns details of fish associated with a specific ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fish details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Peixe"
                }
              }
            }
          },
          "404": {
            "description": "Fish not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UserLogin": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Peixe": {
      "type": "object",
      "required": [
        "ID_usuario",
        "ID_peixes"
      ],
      "properties": {
        "ID_usuario": {
          "type": "string"
        },
        "ID_peixes": {
          "type": "string"
        },
        "Nome": {
          "type": "string"
        }
      }
    },
    "PeixesUsuario": {
      "type": "object",
      "required": [
        "ID_usuario",
        "ID_peixes",
        "Nome"
      ],
      "properties": {
        "ID_usuario": {
          "type": "string"
        },
        "ID_peixes": {
          "type": "string"
        },
        "Nome": {
          "type": "string"
        }
      }
    }
  }
}